pyproject.toml
[project]
name = "artemis-monitor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dsn-parser",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.2",
    "rich>=14.3.3",
]

[dependency-groups]
dev = [
    "maturin>=1.12.6",
    "pytest>=9.0.3",
]

[tool.uv.workspace]
members = ["crates/dsn_parser"]

[tool.uv.sources]
dsn-parser = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["artemis"]

[project.scripts]
artemis-monitor = "artemis.main:main"
